backend/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
pdfplumber==0.11.4
sentence-transformers==3.1.1
faiss-cpu==1.8.0.post1
groq==0.11.0
streamlit==1.38.0
requests==2.32.3
